requirements.txt
bcrypt=4.2.1
blinker=1.9.0
certifi=2025.1.31
cffi=1.17.1
charset-normalizer=3.4.1
click=8.1.8
colorama=0.4.6
cryptography=44.0.0
dnspython=2.7.0
Flask=3.1.0
Flask-Bcrypt=1.0.1
Flask-Cors=5.0.0
Flask-JWT-Extended=4.7.1
idna=3.10
itsdangerous=2.2.0
Jinja2=3.1.5
MarkupSafe=3.0.2
pip=25.0
pycparser=2.22
PyJWT=2.10.1
pymongo=4.11
python-dotenv=1.0.1
requests=2.32.3
setuptools=65.5.0
urllib3=2.3.0
Werkzeug=3.1.3